Added logic for orders

+++ app/models/orders.py
@@ -27,7 +27,7 @@ class Order(Base):
     id: Mapped[int] = mapped_column(Integer, primary_key=True)
     user_id: Mapped[int] = mapped_column(ForeignKey("users.id", ondelete="CASCADE"), nullable=False, index=True)
     status: Mapped[str] = mapped_column(String(20), default="pending", nullable=False)
-    total_amount: Mapped[int] = mapped_column(Numeric(10, 2), server_default="0.00", nullable=False)
+    total_amount: Mapped[Decimal] = mapped_column(Numeric(10, 2), server_default="0.00", nullable=False)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now(), nullable=False)
     updated_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now(), 
                                                  server_onupdate=func.now(), nullable=False)
